Refactored some code out into file utils
